refactored imports of shared schemas
