+++ dynatrace/configuration_v1/oneagent_in_a_hostgroup.py
@@ -20,8 +20,7 @@ from typing import Dict, Any, Optional, Union
 
 from dynatrace.http_client import HttpClient
 from dynatrace.dynatrace_object import DynatraceObject
-from dynatrace.environment_v2.schemas import ConfigurationMetadata
-from dynatrace.configuration_v1.schemas import AutoUpdateSetting, UpdateWindowsConfig, EffectiveSetting
+from dynatrace.configuration_v1.schemas import ConfigurationMetadata, AutoUpdateSetting, UpdateWindowsConfig, EffectiveSetting
 
 
 class OneAgentInAHostGroupService:
