+++ dynatrace/configuration_v1/oneagent_on_a_host.py
@@ -21,8 +21,7 @@ from typing import Dict, Any, Optional
 
 from dynatrace.dynatrace_object import DynatraceObject
 from dynatrace.http_client import HttpClient
-from dynatrace.environment_v2.schemas import ConfigurationMetadata
-from dynatrace.configuration_v1.schemas import AutoUpdateSetting, UpdateWindowsConfig, EffectiveSetting, TechMonitoringList
+from dynatrace.configuration_v1.schemas import ConfigurationMetadata, AutoUpdateSetting, UpdateWindowsConfig, EffectiveSetting, TechMonitoringList
 
 
 class OneAgentOnAHostService:
